fix translation modal team
