+++ src/js/translation/modalTeam.js
@@ -6,20 +6,12 @@ const modalTeamEng =
 const modalTeamUk =
     '<div class="team-cast-box"><div class="team-info-box"><div class="logo-team">logo team</div><p class="team-slogan">Або JavaScript нас, або ми його</p><h1 class="team-name"><a href="https://github.com/byht1/filmoteka-team-project">Назва команди</a></h1><ul class="team-info-list"><li class="team-info-list-item">         <p class="info-caption">Оцінка</p><p class="info-description"><span class="span-orange">10</span></p></li>        <li class="team-info-list-item"><p class="info-caption">Мова оригіналу</p><p class="info-description">HTML, CSS, JavaScript</p></li><li class="team-info-list-item"><p class="info-caption">Країна-виробник</p><p class="info-description">Україна</p></li><li class="team-info-list-item"><p class="info-caption">Жанр</p>         <p class="info-description">Триллер, пригоди, комедія</p></li><li class="team-info-list-item"><p class="info-caption">Дата виходу</p><p class="info-description">2022</p></li><li class="team-info-list-item">     <p class="info-caption">Тривалість</p><p class="info-description">1 тиждень</p></li></ul></div><div class="cast-box"><h2 class="cast-title">Над фільмом працювали</h2><div><ul class="cast-list"><li class="cast-list-item"><p class="cast-role">Продюсер / Team lead</p><p class="cast-name"><a href="https://github.com/byht1">Віталій Дихал</a></p></li><li class="cast-list-item"><p class="cast-role">Режисер / Scrum Master</p><p class="cast-name"><a href="https://github.com/hellene-mary">Марія Грек</a></p></li><li class="cast-list-item"><p class="cast-role">Оператор-постановник / Розробник</p><p class="cast-name"><a href="https://github.com/Veronika-chenko">Veronika Kravchenko</a></p></li><li class="cast-list-item"><p class="cast-role">Художник-постановник / Розробник</p><p class="cast-name"><a href="https://github.com/ViktoriaZinyak">Зиняк Виктория</a></p></li><li class="cast-list-item"><p class="cast-role">Монтажер / Розробник</p><p class="cast-name"><a href="https://github.com/AlexKekc">Джонсонюк</a></p></li><li class="cast-list-item"><p class="cast-role">Кастинг-директор / Розробник</p><p class="cast-name"><a href="https://github.com/Nazarii7">Nazarii Pushkaruk</a></p></li><li class="cast-list-item"><p class="cast-role">Помічник режисера / Розробник</p><p class="cast-name"><a href="https://github.com/zlatan-dp">Олександр Шубін</a></p></li><li class="cast-list-item"><p class="cast-role">Постановник трюків / Розробник</p><p class="cast-name"><a href="https://github.com/Dmytro80">Дмитро Вороневський</a></p></li><li class="cast-list-item"><p class="cast-role">Каскадер / Розробник</p><p class="cast-name"><a href="https://github.com/yanikr">Yanik Rikhter</a></p></li><li class="cast-list-item"><p class="cast-role">Лінійний продюсер / Розробник</p><p class="cast-name"><a href="https://github.com/Lyud0k">Люда</a></p></li></ul></div></div></div><div class="about-box"><h2 class="about-title">About</h2><p class="about-text">Колись зорі склались так, що ці люди зібрались разом, щоб створити щось надзвичайне. Кожен з них має унікальні здібності, нерозкриті таланти та надзвичайні амбіції. Це їхня перша пригода як команди. І саме ти, глядачу, маєш унікальну можливість це побачити!</p></div>';
 
-refs.changeLanguage.addEventListener('input', changeLanguage);
-
-function changeLanguage(evt) {
-    // true - ukr
-    // fasle - eng
-    console.log('~ evt', evt.target.checked);
+export function changeLanguageModalTeam() {
     refs.modalTeamLanguage.innerHTML = '';
 
-    if (evt.target.checked) {
+    if (refs.html.getAttribute('lang') == 'uk') {
         refs.modalTeamLanguage.innerHTML = modalTeamUk;
-
-        refs.html.setAttribute('lang', 'uk');
     } else {
         refs.modalTeamLanguage.innerHTML = modalTeamEng;
-        refs.html.setAttribute('lang', 'en');
     }
 }
