+++ src/partials/modalTeam.html
@@ -10,31 +10,25 @@
             <div class="team-cast-box">
                 <div class="team-info-box">
                     <div class="logo-team">logo team</div>
-                    <!-- Або JavaScript нас, або ми його -->
                     <p class="team-slogan">For the sake of future programming!</p>
                     <h1 class="team-name">
                         <a href="https://github.com/byht1/filmoteka-team-project">Назва команди</a>
                     </h1>
                     <ul class="team-info-list">
                         <li class="team-info-list-item">
-                            <!-- Оцінка -->
                             <p class="info-caption">Vote</p>
                             <p class="info-description"><span class="span-orange">10</span></p>
                         </li>
                         <li class="team-info-list-item">
-                            <!-- <p class="info-caption">Мова оригіналу</p> -->
                             <p class="info-caption">Original Language</p>
                             <p class="info-description">HTML, CSS, JavaScript</p>
                         </li>
                         <li class="team-info-list-item">
-                            <!-- <p class="info-caption">Країна-виробник</p> -->
                             <p class="info-caption">Country</p>
                             <p class="info-description">Ukraine</p>
                         </li>
                         <li class="team-info-list-item">
-                            <!-- Жанр -->
                             <p class="info-caption">Genre</p>
-                            <!-- Триллер, пригоди, комедія -->
                             <p class="info-description">Thriller, adventure, comedy</p>
                         </li>
                         <li class="team-info-list-item">
@@ -42,7 +36,6 @@
                             <p class="info-description">2022</p>
                         </li>
                         <li class="team-info-list-item">
-                            <!-- <p class="info-caption">Тривалість</p> -->
                             <p class="info-caption">Duration</p>
                             <p class="info-description">1 week</p>
                         </li>
@@ -50,12 +43,10 @@
                 </div>
 
                 <div class="cast-box">
-                    <!-- <h2 class="cast-title">Над фільмом працювали</h2> -->
                     <h2 class="cast-title">Production by</h2>
                     <div>
                         <ul class="cast-list">
                             <li class="cast-list-item">
-                                <!-- <p class="cast-role">Продюсер</p> -->
                                 <p class="cast-role">Producer / Team lead</p>
                                 <p class="cast-name">
                                     <a href="https://github.com/byht1">Віталій Дихал</a>
@@ -122,10 +113,6 @@
 
             <div class="about-box">
                 <h2 class="about-title">About</h2>
-                <!-- <p class="about-text">
-                Колись зорі склались так, що ці люди зібрались разом, щоб створити щось надзвичайне. Кожен з них має унікальні здібності, нерозкриті таланти та надзвичайні амбіції. Це їхня перша пригода як команди. І саме ти, глядачу, маєш унікальну можливість це побачити!
-            </p> -->
-
                 <p class="about-text">
                     Once upon a time, the stars aligned in such a way that these people came together to create
                     something extraordinary. Each of them has unique abilities, undiscovered talents and extraordinary
